requirements.txt
fastapi
uvicorn
pydub
pymongo
funasr
modelscope
torch
torchaudio
python-dotenv
requests
python-jose
python-multipart
qiniu
loguru
langchain
openai
python-dotenv
websockets
